Fix GTK rendering of disabled menu components.
This file has not been modified since the initial OpenJDK import ~10 years ago, so it is not known for sure why this special case was included, but it was likely a workaround to get things rendering properly at the time. This code however causes very poor looking text on dark menu backgrounds (https://bugs.launchpad.net/bugs/1729558), but removing it results in proper rendering both on light and dark backgrounds. The code below it references some other special cases for the GTK "blueprint" theme which is/was used in OpenSolaris. It is unknown whether this code is still required (the blueprint theme seems all but abandoned), but it has been kept to avoid any unexpected changes in behaviour.